move getrackdata server action to be server util
